% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSample.R
\name{addSample}
\alias{addSample}
\title{Add sample interactively}
\usage{
addSample(source_df = sourceDataFrame(), update)
}
\arguments{
\item{source_df}{The source data.frame. Defaults to the current \code{source_df}
of the package.}

\item{update}{Logical. Use with caution! If TRUE, the updated output data.frame 
will replace the current \code{source_df} of the package!}
}
\value{
The updated data.frame is returned invisibly.
}
\description{
Adds a new sample to the source data.frame.
}
