% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{adjustSampleManually}
\alias{adjustSampleManually}
\title{Adjust sample info manually}
\usage{
adjustSampleManually(
  source_df = sourceDataFrame(),
  sample_name,
  adjust_sample_input,
  update
)
}
\arguments{
\item{source_df}{The source data.frame. Defaults to the current \code{source_df}
of the package.}

\item{sample_name}{Character value. The sample that is supposed to 
be adjusted.}

\item{adjust_sample_input}{A named list. Each slot 
corresponds to the variable of \code{source_df} that is supposed
to be adjusted.}

\item{update}{Logical. Use with caution! If TRUE, the updated output data.frame 
will replace the current \code{source_df} of the package!}
}
\value{
The updated data.frame is returned invisibly.
}
\description{
Instead of adding a new sample the samples info can be 
manually adjusted using this function.
}
\examples{

 # adjust image link
 
 adjustSampleManually(
    source_df = sourceDataFrame(), # default 
    sample_name = "275_T", 
    adjust_sample_input = list(link_image = "https://some-link-to-image.jpeg?dl=1")
    update = TRUE # immediately saves the result in /data/source_df.rda
    )

}
