% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{argument_dummy}
\alias{argument_dummy}
\title{Argument dummy}
\usage{
argument_dummy()
}
\arguments{
\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Argument dummy
}
\keyword{internal}
