% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadRawData}
\alias{downloadRawData}
\title{Download raw data}
\usage{
downloadRawData()

downloadRawData()
}
\arguments{
\item{sample_names}{Character vector. The sample names of the data
to be downloaded. Use \code{validSampleNames()} to obtain all valid input options.}
}
\value{
An invisible TRUE.
}
\description{
Downloads raw Visium10X output of samples as a .zip folder.
}
\details{
The directory of the .zip folder is eventually created
with \code{stringr::str_c(folder, "/", sample, ".zip")}. To access
the data you have to unzip the folder manually.
}
\keyword{internal}
