% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kueckelj.R
\name{downloadSpataObjects}
\alias{downloadSpataObjects}
\title{Download several spata objects}
\usage{
downloadSpataObjects(
  sample_names,
  files = NULL,
  folder = base::getwd(),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sample_names}{Character vector. The sample names of the spata objects
to be downloaded. Use \code{validSampleNames()} to obtain all valid input options.}

\item{files}{Character vector or NULL. Specifies the file names under which the
spata objects are saved. If character, the input must be of the same length
as the input for argument \code{sample_names}. If NULL, the files are named
according to the sample name.}

\item{folder}{Character value. If character, specifies the output
folder in which the spata object is saved. Defaults to the working directory.}

\item{overwrite}{Logical. Must be set to TRUE if file directories
under which downloaded files are about to be saved already exist.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An invisible TRUE if everything worked flawlessly.
}
\description{
Main function that downloads several spata objects
at the same time and saves them as an .RDS file.
}
\examples{

# downloads three spata objects and
# stores them as "spata_data/275_T.RDS", "spata_data/313_T.RDS" etc.

  downloadSpataObjects(
    sample_names = ("275_T", "313_T", "334_T"),
    folder = "spata_data"
   )

}
