% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{features_list}
\alias{features_list}
\title{Meta data of features and cell type deconvolution}
\format{
An object of class \code{list} of length 16.
}
\usage{
features_list
}
\description{
A list of data.frames named by samples of the samples available. 
The cell type annotations were performed by spaceXR 
Cable, D.M., Murray, E., Shanmugam, V. et al. 
Cell type-specific inference of differential expression in 
spatial transcriptomics. Nat Methods (2022). 
https://doi.org/10.1038/s41592-022-01575-3 
using the reference scRNA-seq dataset from Cristian Ruiz-Moreno 
"https://www.biorxiv.org/content/10.1101/2022.08.27.505439v1" (not published)
}
\keyword{datasets}
