% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getCitation}
\alias{getCitation}
\alias{getCitationBySample}
\title{Obtain citation info}
\usage{
getCitation(object, write_lines = TRUE)

getCitationBySample(sample_names)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{write_lines}{Logical. If \code{TRUE}, the string is printed in the
console via \code{writeLines()}.}

\item{sample_names}{Character vector of sample names for which
you want to obtain the correct citation.}
}
\value{
Citation and - if one of SPATAData objects - the download source.
}
\description{
If you have used a downloadable spata object
please use this function to obtain the proper
form of citation to give credits to the researchers who
made this data available
}
