% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSampleManually.R
\name{removeSampleManually}
\alias{removeSampleManually}
\title{Remove sample}
\usage{
removeSampleManually(sample_name, update)
}
\arguments{
\item{sample_name}{The sample name(s) to be removed.}

\item{update}{Logical. Use with caution! If TRUE, the updated output data.frame 
will replace the current \code{source_df} of the package!}
}
\value{
The updated data.frame is returned invisibly.
}
\description{
Removes a sample from the \code{source_df} of the package.
}
\details{
As a precaution \code{udpate} must be manually set to TRUE.
}
