% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{sourceDataFrameTags}
\alias{sourceDataFrameTags}
\alias{sampleTags}
\title{Obtain all sample tags}
\usage{
sourceDataFrameTags(source_df = sourceDataFrame())

sampleTags(source_df = sourceDataFrame(), sample_name)
}
\arguments{
\item{sample_name}{Character value. The sample of interest.}
}
\value{
Character vector.
}
\description{
Returns sample tags of the \code{source_df}.
}
