% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{validSampleNames}
\alias{validSampleNames}
\title{Valid sample names}
\usage{
validSampleNames()
}
\value{
Character vector.
}
\description{
Returns the sample names of the samples that you can download
via \code{downloadSpataObject()}, \code{downloadSpataObject(()}.
}
