% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_kueckelj.R
\name{downloadRawData}
\alias{downloadRawData}
\title{Download raw data}
\usage{
downloadRawData(
  sample_names,
  folder = base::getwd(),
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{sample_names}{Character vector. The sample names of the data
to be downloaded. Use \code{validSampleNames()} to obtain all valid input options.}

\item{folder}{Character value. If character, specifies the output
folder in which the spata object is saved. Defaults to the working directory.}

\item{overwrite}{Logical. Must be set to TRUE if file directories
under which downloaded files are about to be saved already exist.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An invisible TRUE.
}
\description{
Downloads raw Visium10X output of samples as a .zip folder.
}
\details{
The directory of the .zip folder is eventually created
with \code{stringr::str_c(folder, "/", sample, ".zip")}. To access
the data you have to unzip the folder manually.
}
