% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadSpataObject}
\alias{downloadSpataObject}
\title{Download spata objects}
\usage{
downloadSpataObject(
  sample_name,
  overwrite = FALSE,
  folder = base::getwd(),
  file = NULL,
  in_shiny = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sample_name}{Character value. The name of the sample you want to
download. Use \code{validSampleNames()} to obtain all valid input options.}

\item{overwrite}{Logical. Must be set to TRUE if file directories
under which downloaded files are about to be saved already exist.}

\item{folder}{Character value. If character, specifies the output
folder in which the spata object is saved. Defaults to the working directory.}

\item{file}{The filename of the spata object. Must end with \emph{'.RDS'}. By
default the file it is `NULL` which makes the function skip the saving step.
Set to character, if you want the object to be saved.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The downloaded spata object.
}
\description{
Downloads a spata object and returns it. For convenient
downloads of multiple spata objects check out \code{downloadSpataObjects()}.
}
\details{
The downloaded spata object is immediately saved after the download before
it is returned by the function.
}
\examples{

# only download
downloadSpataObject(sample_name = "275_T", folder = getwd())

# download AND assign
object <- downloadSpataObject(sample_name = "275_T", folder = getwd())

# only assign (default)
object <- downloadSpataObject(sample_name = "275_T", file = NULL)


}
