% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_jk.R
\name{downloadSpataObject}
\alias{downloadSpataObject}
\title{Download spata objects}
\usage{
downloadSpataObject(
  sample_name,
  overwrite = FALSE,
  folder = base::getwd(),
  file = stringr::str_c(sample_name, ".RDS"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sample_name}{Character value. The name of the sample you want to
download. Use \code{validSampleNames()} to obtain all valid input options.}

\item{overwrite}{Logical. Must be set to TRUE if file directories
under which downloaded files are about to be saved already exist.}

\item{folder}{Character value. If character, specifies the output
folder in which the spata object is saved. Defaults to the working directory.}

\item{file}{The filename of the spata object. Must end with \emph{'.RDS'}. By
default the file is named like the sample name. Set to NULL if you don't
want the object to be saved.}
}
\value{
The downloaded spata object.
}
\description{
Downloads spata objects from the internet and saves them
on your device. See details for more information about the differences
between \code{downloadSpataObject()} and \code{downloadSpataObjects()}.
}
\details{
The downloaded spata object is immediately saved after the download before
it is returned by the function.
}
\examples{

# only download
downloadSpataObject(sample_name = "275_T")

# download AND assign
object <- downloadSpataObject(sample_name = "275_T")


}
