% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{downloadSpataObject}
\alias{downloadSpataObject}
\title{Download objects of class \code{SPATA2}}
\usage{
downloadSpataObject(
  sample_name,
  overwrite = FALSE,
  file = FALSE,
  in_shiny = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sample_name}{Character value. The name of the sample you want to
download. Use \code{validSampleNames()} to obtain all valid input options.}

\item{overwrite}{Logical. Must be set to TRUE if file directories
under which downloaded files are about to be saved already exist.}

\item{file}{The filename of the \code{SPATA2} object. Must end with \emph{'.RDS'}. By
default the file it is \code{NULL} which makes the function save the object
under the sample name with an \emph{'.RDS'} suffix. If \code{FALSE}, the saving
is skipped and the object is simply returned.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{folder}{Character value. If character, specifies the output
folder in which the \code{SPATA2} object is saved. Defaults to the working directory.}
}
\value{
The downloaded \code{SPATA2} object.
}
\description{
Downloads a \code{SPATA2} object and returns it. For convenient
downloads of multiple \code{SPATA2} objects check out \code{downloadSpataObjects()}.
}
\details{
If \code{file} is not \code{FALSE}. The downloaded \code{SPATA2} object is immediately saved after the download before
it is returned by the function. Note that the file directory is assembled by combining
\code{folder} and \code{file}!
}
\examples{

# download & assign (no saving on the disk)
object <- downloadSpataObject(sample_name = "UKF275T")

# download, assign and save on disk
# -> stores the file under ~/UKF275T.RDS (where '~' is your working directory)
object <- downloadSpataObject(sample_name = "UKF275T", file = TRUE)

# download, assign and save on disk in a specified directory
object <- downloadSpataObject(sample_name = "UKF275T", file = "my/path/spata_object.RDS")

}
