% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_jk.R
\name{getCitation}
\alias{getCitation}
\alias{getCitationBySample}
\title{Obtain citation info}
\usage{
getCitation(object)

getCitationBySample(sample_names = validSampleNames())
}
\arguments{
\item{sample_names}{Character vector of sample names for which
you want to obtain the correct citation.}
}
\value{
Citation in form of character vectors or a list of such.
}
\description{
If you have used a downloadable spata object
please use this function to obtain the proper
form of citation to give credits to the researchers who
made this data available
}
