% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(sample_name, folder = NULL, type = "SPATA")
}
\arguments{
\item{type}{Can be either 'SPATA' to download the Spata objects or 'RAW' for access the raw data}
}
\value{

}
\description{
getData
}
\examples{

}
\author{
Dieter Henrik Heiland
}
