% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_jk.R
\name{setCitation}
\alias{setCitation}
\title{Set citation info}
\usage{
setCitation(object, citation)
}
\arguments{
\item{citation}{Character value.}
}
\value{
An updated spata object.
}
\description{
Sets information about how to cite this
spata object.
}
