% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{setCitation}
\alias{setCitation}
\title{Set citation info}
\usage{
setCitation(object, citation)
}
\arguments{
\item{object}{An object of class \code{\link[SPATA2]{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{citation}{Character value.}
}
\value{
An updated spata object.
}
\description{
Sets information about how to cite this
spata object.
}
