% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{setCitation}
\alias{setCitation}
\title{Set citation info}
\usage{
setCitation(object, citation)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{citation}{Character value containing the complete citation.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Sets information about how to cite this
\code{SPATA2} object.
}
\details{
Input for \code{citation} is set in @meta_sample$pub_citation.
}
