% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{source_df}
\alias{source_df}
\title{Meta data and links for MILO Lab database}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6 rows and 21 columns.
}
\usage{
source_df
}
\description{
A data.frame of multiple variables containing information 
around the data sets that are available in \code{SPATAData}.
}
\keyword{datasets}
