% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{sourceDataFrame}
\alias{sourceDataFrame}
\title{Obtain the SPATAData source data.frame}
\usage{
sourceDataFrame()
}
\value{
Data.frame in which each row corresponds to a spatial data set stored
in a \code{SPATA2} object.
}
\description{
Returns the \link[=source_df]{source data.frame}.
}
\examples{

library(SPATA2)
library(SPATAData)
library(dplyr)
library(stringr)

sdf <- sourceDataFrame()

#----- example dplyr logic to filter the source data.frame

# 1. obtain glioblastoma samples from the temporal lobe

sdf_gbm <- filter(sdf, histo_class == "Glioblastoma" & organ_part == "temporal")

gbm_samples <- sdf_gbm$sample_name

print(gbm_samples)

# downlaod as collection
downloadSpataObjects(sample_names = gbm_samples, folder = "spata_objects/gbm") 

# 2. obtain data from specific publications

sdf_kuppe <- 
 filter(sdf, str_detect(pub_citation, pattern = "^Kuppe") & pub_journal == "Nature")
 
kuppe_samples <- sdf_kuppe$sample_name

print(kuppe_samples)

}
