% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{sourceDataFrame}
\alias{sourceDataFrame}
\title{Obtain the SPATAData \code{source_df}}
\usage{
sourceDataFrame(organ = NULL, status = NULL)
}
\arguments{
\item{organ}{Character value. Filter for organs.}

\item{status}{Character value. Filter for tissue status. \emph{'p'} equals
pathological. \emph{'h'} equals healthy.}
}
\value{
Data.frame.
}
\description{
Returns the \code{source_df} and allows some filtering.
}
