% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{sourceDataFrame}
\alias{sourceDataFrame}
\title{SPATAData source data.frame}
\usage{
sourceDataFrame()
}
\value{
Data.frame in which each row corresponds to a spatial data set stored
in a \code{SPATA2} object.

\itemize{
\item{sample_name}{: Character. Name of the sample and its unique identifier.}
\item{comment}{: Character. Additional comments about the sample.}
\item{donor_id}{: Character. Unique identifier for the donor.}
\item{donor_species}{: Character. Species of the donor.}
\item{grade}{: Character. Grade of the sample.}
\item{grade_sub}{: Character. Sub-grade of the sample.}
\item{histo_class}{: Character. Histological classification.}
\item{histo_class_sub}{: Character. Sub-classification of the histological class.}
\item{institution}{: Character. Institution where the sample was collected.}
\item{lm_source}{: Date-time. Last instance when the row of the source data.frame was modified.}
\item{organ}{: Character. Organ from which the sample was taken.}
\item{organ_part}{: Character. Specific part of the organ from which the sample was taken.}
\item{pathology}{: Character. Pathological state of the sample.}
\item{platform}{: Character. Platform used for the experiment.}
\item{pub_citation}{: Character. Citation for the publication related to the sample.}
\item{pub_doi}{: Character. DOI of the publication related to the sample.}
\item{pub_journal}{: Character. Journal where the related publication was published.}
\item{pub_year}{: Numeric. Year of publication.}
\item{sex}{: Character. Sex of the donor.}
\item{side}{: Character. Side of the organ from which the sample was taken.}
\item{tags}{: Character. Tags related to the sample.}
\item{tissue_age}{: Numeric. Age of the tissue in years.}
\item{web_link}{: Character. Link to an object or additional data related to the sample for downloads.}
\item{workgroup}{: Character. Workgroup or team responsible for the sample.}
\item{mean_counts}{: Numeric. Mean counts of the measurements.}
\item{median_counts}{: Numeric. Median counts of the measurements.}
\item{modality_gene}{: Logical. Indicates if the modality includes genes.}
\item{modality_metabolite}{: Logical. Indicates if the modality includes metabolites.}
\item{modality_protein}{: Logical. Indicates if the modality includes proteins.}
\item{n_obs}{: Numeric. Number of observations.}
\item{n_tissue_sections}{: Numeric. Number of tissue sections.}
\item{obs_unit}{: Character. Unit of observation.}
}
}
\description{
Access the source data.frame which contains information about
various spatial transcriptomic experiments .
}
\examples{

library(SPATA2)
library(SPATAData)
library(dplyr)
library(stringr)

sdf <- sourceDataFrame()

#----- example dplyr logic to filter the source data.frame

# 1. obtain glioblastoma samples from the temporal lobe

sdf_gbm <- filter(sdf, histo_class == "Glioblastoma" & organ_part == "temporal")

gbm_samples <- sdf_gbm$sample_name

print(gbm_samples)

# downlaod as collection
downloadSpataObjects(sample_names = gbm_samples, folder = "spata_objects/gbm") 

# 2. obtain data from specific publications

sdf_kuppe <- 
 filter(sdf, str_detect(pub_citation, pattern = "^Kuppe") & pub_journal == "Nature")
 
kuppe_samples <- sdf_kuppe$sample_name

print(kuppe_samples)

}
