% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_jk.R
\name{validSampleNames}
\alias{validSampleNames}
\title{Valid spata object names}
\usage{
validSampleNames()
}
\value{
Character vector.
}
\description{
Returns the sample names
of the spata objects that you can download
via \code{downloadSpataObject()} and \code{downloadSpataObject()}.
}
